pyproject.toml
[tool.poetry]
name = "bibli"
version = "0.1.0"
description = ""
authors = ["Corentin Dancette <corentin.dancette@gmail.com>"]
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
flask = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
